Move classes to logical packages
