+++ src/test/java/nl/tudelft/contextproject/util/webinterface/EntityUtilTest.java
@@ -1,5 +1,6 @@
-package nl.tudelft.contextproject.util;
+package nl.tudelft.contextproject.util.webinterface;
 
+import nl.tudelft.contextproject.util.webinterface.EntityUtil;
 import org.junit.Test;
 
 import nl.tudelft.contextproject.TestBase;
